dev/releases/utils.py: move GitHub specific code to new file
... so that one can use the utils code without having the "github"
python package installed.
